feat(route): 机核网添加分类标签和标签路由 (#9239)
* Gcores tagging

* Gcore tags

* Comfort v2

* Update router

* Update router.js

* refactor: use cache.tryGet
