+++ lib/router.js
@@ -338,9 +338,6 @@ router.get('/t66y/:id/:type?', lazyloadRouteHandler('./routes/t66y/index'));
 // 色中色
 router.get('/sexinsex/:id/:type?', lazyloadRouteHandler('./routes/sexinsex/index'));
 
-// 机核
-router.get('/gcores/category/:category', lazyloadRouteHandler('./routes/gcores/category'));
-
 // 国家地理
 router.get('/natgeo/dailyphoto', lazyloadRouteHandler('./routes/natgeo/dailyphoto'));
 router.get('/natgeo/:cat/:type?', lazyloadRouteHandler('./routes/natgeo/natgeo'));
