+++ lib/v2/gcores/maintainer.js
@@ -1,0 +1,4 @@
+module.exports = {
+    '/category/:category': ['MoguCloud', 'StevenRCE0'],
+    '/tag/:tag/:category?': ['StevenRCE0'],
+};
