+++ lib/v2/gcores/router.js
@@ -1,0 +1,4 @@
+module.exports = function (router) {
+    router.get('/category/:category', require('./category'));
+    router.get('/tag/:tag/:category?', require('./tag'));
+};
